Add files via upload
Skeletal code for running the stm. I prefixed the file with a number so that there is some logic to how files are stored in the folder. 

FYI, I kept receiving a weird "unknown input format" error when trying to read the WaterPlans_text.rds file. It might be a setup issue on my end so will try on my laptop at home tonight as well.